NEW/test_final_comprehensive.py
#!/usr/bin/env python3
"""
Final Comprehensive Test for ClauseWise with pyttsx3 TTS

This script performs a complete test of all ClauseWise functionality
to ensure everything is working correctly before final deployment.
"""

import sys
import os
import tempfile

def test_dependencies():
    """Test all required dependencies"""
    print("🔍 Testing Dependencies")
    print("=" * 23)
    
    dependencies = [
        ("streamlit", "Streamlit web framework"),
        ("requests", "HTTP requests library"),
        ("pdfplumber", "PDF text extraction"),
        ("docx", "DOCX document processing"),
        ("plotly", "Data visualization"),
        ("pandas", "Data manipulation"),
        ("numpy", "Numerical computing"),
        ("pyttsx3", "Offline text-to-speech")
    ]
    
    success_count = 0
    for dep, description in dependencies:
        try:
            if dep == "docx":
                import docx
            else:
                __import__(dep)
            print(f"✅ {dep}: {description}")
            success_count += 1
        except ImportError:
            print(f"❌ {dep}: {description} - NOT INSTALLED")
    
    print(f"\n📊 Dependencies: {success_count}/{len(dependencies)} available")
    return success_count == len(dependencies)

def test_pyttsx3_functionality():
    """Test pyttsx3 TTS functionality"""
    print("\n🔍 Testing pyttsx3 TTS Functionality")
    print("=" * 35)
    
    try:
        import pyttsx3
        
        # Test engine initialization
        engine = pyttsx3.init()
        print("✅ pyttsx3 engine initialized")
        
        # Test voice properties
        voices = engine.getProperty('voices')
        if voices and len(voices) > 0:
            print(f"✅ Found {len(voices)} voices available")
        else:
            print("⚠️ No voices found, but engine works")
        
        # Test basic speech generation
        test_text = "This is a test of the offline text to speech system."
        
        with tempfile.NamedTemporaryFile(delete=False, suffix='.wav') as temp_file:
            temp_path = temp_file.name
        
        try:
            engine.save_to_file(test_text, temp_path)
            engine.runAndWait()
            
            if os.path.exists(temp_path) and os.path.getsize(temp_path) > 0:
                print(f"✅ Audio file generated: {os.path.getsize(temp_path)} bytes")
                os.unlink(temp_path)
                engine.stop()
                return True
            else:
                print("❌ Audio file generation failed")
                if os.path.exists(temp_path):
                    os.unlink(temp_path)
                engine.stop()
                return False
                
        except Exception as e:
            print(f"❌ Audio generation error: {e}")
            if os.path.exists(temp_path):
                os.unlink(temp_path)
            engine.stop()
            return False
            
    except Exception as e:
        print(f"❌ pyttsx3 test failed: {e}")
        return False

def test_core_functions():
    """Test core ClauseWise functions"""
    print("\n🔍 Testing Core ClauseWise Functions")
    print("=" * 36)
    
    # Mock streamlit for testing
    class MockStreamlit:
        def info(self, msg): pass
        def success(self, msg): pass
        def warning(self, msg): pass
        def error(self, msg): pass
    
    sys.modules['streamlit'] = MockStreamlit()
    
    try:
        # Test legal term replacement
        test_text = "Whereas the party hereby covenants to indemnify the other party."
        
        # Simple replacement test
        replacements = {
            "whereas": "while",
            "hereby": "by this document",
            "covenant": "promise",
            "indemnify": "protect from loss"
        }
        
        simplified = test_text.lower()
        for legal_term, plain_term in replacements.items():
            simplified = simplified.replace(legal_term, plain_term)
        
        improvements = sum(1 for term in replacements.keys() if term not in simplified)
        print(f"✅ Legal term replacement: {improvements}/{len(replacements)} terms simplified")
        
        # Test entity extraction patterns
        import re
        sample_text = "This agreement dated March 15, 2024, involves $50,000 between ABC Corp and john@email.com."
        
        dates = re.findall(r'\b(?:January|February|March|April|May|June|July|August|September|October|November|December)\s+\d{1,2},?\s+\d{4}\b', sample_text)
        money = re.findall(r'\$[\d,]+\.?\d*', sample_text)
        emails = re.findall(r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b', sample_text)
        
        entities_found = len(dates) + len(money) + len(emails)
        print(f"✅ Entity extraction: {entities_found} entities found (dates: {len(dates)}, money: {len(money)}, emails: {len(emails)})")
        
        return improvements > 0 and entities_found >= 3
        
    except Exception as e:
        print(f"❌ Core functions test failed: {e}")
        return False

def test_tts_integration():
    """Test TTS integration with ClauseWise"""
    print("\n🔍 Testing TTS Integration")
    print("=" * 26)
    
    # Mock streamlit
    class MockStreamlit:
        def info(self, msg): print(f"INFO: {msg}")
        def success(self, msg): print(f"SUCCESS: {msg}")
        def warning(self, msg): print(f"WARNING: {msg}")
        def error(self, msg): print(f"ERROR: {msg}")
    
    sys.modules['streamlit'] = MockStreamlit()
    
    try:
        # Import TTS function
        sys.path.append(os.path.dirname(os.path.abspath(__file__)))
        from utils import text_to_speech
        
        # Test with sample legal text
        legal_text = "This legal document establishes the terms and conditions for the agreement between the parties."
        
        result = text_to_speech(legal_text)
        
        if result and len(result) > 1000:  # Should be a substantial base64 string
            print(f"✅ TTS integration successful: {len(result)} character audio data")
            return True
        else:
            print("❌ TTS integration failed or returned insufficient data")
            return False
            
    except Exception as e:
        print(f"❌ TTS integration test failed: {e}")
        return False

def test_file_processing():
    """Test file processing capabilities"""
    print("\n🔍 Testing File Processing")
    print("=" * 25)
    
    try:
        # Test text processing
        sample_text = """
        **LEGAL CONTRACT**
        
        This agreement contains •important terms• and *conditions*.
        
        The parties shall comply with all obligations.
        """
        
        # Clean text (similar to TTS preprocessing)
        import re
        clean_text = sample_text.replace("**", "").replace("•", "").replace("*", "")
        clean_text = clean_text.replace("\n", " ").replace("\r", " ")
        clean_text = re.sub(r'[^\w\s.,!?-]', '', clean_text)
        clean_text = " ".join(clean_text.split())
        
        if len(clean_text) > 0 and "LEGAL CONTRACT" in clean_text:
            print("✅ Text preprocessing successful")
            return True
        else:
            print("❌ Text preprocessing failed")
            return False
            
    except Exception as e:
        print(f"❌ File processing test failed: {e}")
        return False

def main():
    """Run comprehensive tests"""
    print("🚀 ClauseWise Final Comprehensive Test")
    print("=" * 38)
    print("Testing all functionality before deployment...\n")
    
    tests = [
        ("Dependencies", test_dependencies),
        ("pyttsx3 TTS", test_pyttsx3_functionality),
        ("Core Functions", test_core_functions),
        ("TTS Integration", test_tts_integration),
        ("File Processing", test_file_processing)
    ]
    
    results = {}
    for test_name, test_func in tests:
        try:
            results[test_name] = test_func()
        except Exception as e:
            print(f"❌ {test_name} test crashed: {str(e)}")
            results[test_name] = False
    
    # Final Summary
    print("\n" + "="*50)
    print("📊 FINAL TEST RESULTS")
    print("="*50)
    
    passed = 0
    critical_passed = 0
    critical_tests = ["Dependencies", "pyttsx3 TTS", "TTS Integration"]
    
    for test_name, success in results.items():
        status = "✅ PASS" if success else "❌ FAIL"
        critical = "🔥 CRITICAL" if test_name in critical_tests else ""
        print(f"   {test_name}: {status} {critical}")
        
        if success:
            passed += 1
            if test_name in critical_tests:
                critical_passed += 1
    
    print(f"\n🎯 Overall Results:")
    print(f"   Total Tests: {passed}/{len(tests)} passed")
    print(f"   Critical Tests: {critical_passed}/{len(critical_tests)} passed")
    
    # Final verdict
    if passed == len(tests):
        print("\n🎉 ALL TESTS PASSED! ClauseWise is ready for deployment!")
        print("✅ pyttsx3 offline TTS working perfectly")
        print("✅ All core functions operational")
        print("✅ Enhanced features active")
        print("✅ Professional UI ready")
        return True
    elif critical_passed == len(critical_tests):
        print("\n⚠️ CRITICAL TESTS PASSED! ClauseWise is ready with minor issues")
        print("✅ pyttsx3 offline TTS working")
        print("✅ Core functionality operational")
        return True
    else:
        print("\n❌ CRITICAL TESTS FAILED! Please fix issues before deployment")
        return False

if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)
